cleanup constants
